# etc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Homer-App User interface for WEB AI",
    "title": "Homer-App",
    "license": {
      "name": "AGPL",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.1.2"
  },
  "host": "localhost:9080",
  "basePath": "/api/v3",
  "paths": {
...
    "/user/profile": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns the list of settings",
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsGetAll",
        "responses": {
          "200": {
            "description": "UserSettingList",
            "schema": {
              "$ref": "#/definitions/UserSettingList"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/user/settings": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns the list of user settings",
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsGetAll",
        "responses": {
          "200": {
            "description": "UserSettingList",
            "schema": {
              "$ref": "#/definitions/UserSettingList"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Add new user setting",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsAddUserSettings",
        "parameters": [
          {
            "type": "object",
            "description": "UserSetting struct",
            "name": "UserSetting",
            "in": "body",
            "required": true,
            "schema": {
              "description": "UserSetting struct",
              "type": "object",
              "$ref": "#/definitions/UserSetting"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "UserSettingCreateSuccessfulResponse",
            "schema": {
              "$ref": "#/definitions/UserSettingCreateSuccessfulResponse"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/user/settings/{category}": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns the list of user settings",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsGetCategory",
        "parameters": [
          {
            "type": "string",
            "description": "user settings category",
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "description": "user settings category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UserSettingList",
            "schema": {
              "$ref": "#/definitions/UserSettingList"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/user/settings/{guid}": {
      "put": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Update user setting",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsUpdateUserSettings",
        "parameters": [
          {
            "type": "string",
            "description": "guid of user settings",
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "description": "guid of user settings",
              "type": "string"
            }
          },
          {
            "type": "object",
            "description": "UserSetting",
            "name": "UserSetting",
            "in": "body",
            "required": true,
            "schema": {
              "description": "UserSetting",
              "type": "object",
              "$ref": "#/definitions/UserSetting"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "UserSettingUpdateSuccessfulResponse",
            "schema": {
              "$ref": "#/definitions/UserSettingUpdateSuccessfulResponse"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Delete user setting",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "operationId": "settingsDeleteUserSettings",
        "parameters": [
          {
            "type": "string",
            "description": "guid of user settings",
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "description": "guid of user settings",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "UserSettingDeleteSuccessfulResponse",
            "schema": {
              "$ref": "#/definitions/UserSettingDeleteSuccessfulResponse"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DashboardElement": {
      "type": "object",
      "properties": {
        "cssclass": {
          "type": "string",
          "x-go-name": "CssClass",
          "example": "fa"
        },
        "href": {
          "type": "string",
          "x-go-name": "Href",
          "example": "_1631547759877"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id",
          "example": "_1631547759877"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "Search"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner",
          "example": "admin"
        },
        "param": {
          "type": "string",
          "x-go-name": "Param",
          "example": "search"
        },
        "shared": {
          "type": "boolean",
          "x-go-name": "Shared",
          "example": false
        },
        "type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Type",
          "example": 1
        },
        "weight": {
          "type": "number",
          "format": "double",
          "x-go-name": "Weight",
          "example": 10
        }
      },
      "x-go-name": "DashBoardElement",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
    "FailureResponse": {
      "type": "object",
      "properties": {
        "error": {
          "description": "errot",
          "type": "string",
          "x-go-name": "Error"
        },
        "message": {
          "description": "message",
          "type": "string",
          "x-go-name": "Message"
        },
        "statuscode": {
          "description": "statuscode",
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusCode"
        }
      },
      "x-go-name": "UserTokenBadResponse",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
    "UserSetting": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "x-go-name": "Category",
          "example": "dashboard"
        },
        "data": {
          "$ref": "#/definitions/DashboardElement"
        },
        "guid": {
          "type": "string",
          "x-go-name": "GUID",
          "example": "1868f318-5f16-4c40-ab1e-bfa4f417ff61"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "example": 59
        },
        "param": {
          "type": "string",
          "x-go-name": "Param",
          "example": "_1633964934444"
        },
        "partid": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Partid",
          "example": 10
        },
        "username": {
          "type": "string",
          "x-go-name": "Username",
          "example": "admin"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "UserSettingCreateSuccessfulResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f4e2953e-ab42-40df-a7de-9ceb7faca396"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully created userObject"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "UserSettingDeleteSuccessfulResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f4e2953e-ab42-40df-a7de-9ceb7faca396"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully deleted userObject"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "UserSettingList": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count",
          "example": 7
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserSetting"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "UserSettingUpdateSuccessfulResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f4e2953e-ab42-40df-a7de-9ceb7faca396"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully updated userObject\""
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}